Rename profiler
